Sorted plots v1

+++ size_sorting.py
@@ -121,6 +121,7 @@ file_names = [r"uninhb_sorted.xlsx",
               r"inhb_sorted.xlsx",
               r"inhb_del_sorted.xlsx",
               r"reim_uninhb_sorted.xlsx"]
+file_name = []
 
 # Set names for the three types of particles
 types = ['S-phase', 'Theta', 'Secondary']
@@ -139,6 +140,9 @@ df = np.random.randn(1, 1)
 data = pd.DataFrame(df)
 for file in file_names:
     file_dir = os.path.join(new_folder_dir, file)
+    if not os.path.exists(file_dir):
+        # print("Error:", file, "does not exist. Skipping.")
+        continue
     writer = pd.ExcelWriter(file_dir, engine='openpyxl', mode='w')
     data.to_excel(writer, sheet_name="All")
     writer.close()
@@ -146,6 +150,9 @@ for file in file_names:
 # PREPARE ORIGINAL FILES
 for file in file_dirs:
     file_dir = os.path.join(folder_dir, file)
+    if not os.path.exists(file_dir):
+        print("Error:", file, "does not exist. Skipping.")
+        continue
     data = pd.read_excel(file_dir, "Sheet1", header=None)
     data.iloc[0, 0] = "ROI"
     if str(data.iloc[1, 0]).upper() == "NAN":
@@ -159,7 +166,12 @@ for file in file_dirs:
 for file in file_dirs:
     # Set file directories
     file_dir = os.path.join(folder_dir, file)
-    new_file_dir = os.path.join(new_folder_dir, file_names[i])
+    if not os.path.exists(file_dir):
+        file_dirs.remove(file)
+        # print("Error:", file, "does not exist. Skipping.")
+        continue
+    file_name.append(file_names[i])
+    new_file_dir = os.path.join(new_folder_dir, file_name[i])
     # Read only the wanted columns (ROI, Area, Type and geometry parameter if specified)
     data = pd.read_excel(file_dir, "Sheet1", usecols=cols)
     # Write all data to the new file
@@ -236,7 +248,10 @@ i = 0
 with open(os.path.join(new_folder_dir, "pre-processing summary.txt"), 'w') as f:
     # Print selected options
     f.write("Chosen geometry parameter: "+str(geometry_parameter)+"\nChosen filter mode: "+str(mode_selection)+"\n\n")
-    for file in file_names:
+    for file in file_name:
+        test = pd.read_excel(os.path.join(new_folder_dir, file), sheet_name=None)
+        if "S-phase" not in test.keys():
+            continue
         # Print the test type
         data = pd.read_excel(os.path.join(new_folder_dir, file), sheet_name=sheets)
         f.write("\n"+sets[i]+":")
